Update: Fri

+++ magick/quantize.c
@@ -1938,7 +1938,8 @@ unsigned int QuantizationError(Image *image)
   /*
     Compute final error statistics.
   */
-  image->mean_error_per_pixel=total_error/(image->columns*image->rows);
+  image->mean_error_per_pixel=(unsigned int)
+    (total_error/(image->columns*image->rows));
   image->normalized_mean_error=
     (image->mean_error_per_pixel)/(3.0*(MaxRGB+1)*(MaxRGB+1));
   image->normalized_maximum_error=
